requirements.txt
anyio==4.11.0
APScheduler==3.11.0
certifi==2025.10.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
python-telegram-bot==20.5
sniffio==1.3.1
typing_extensions==4.15.0
tzlocal==5.3.1
